feat: 지난 시간 예약 요청 예외 처리 추가

+++ src/main/java/roomescape/theme/application/ThemeService.java
@@ -6,6 +6,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.domain.ReservationRepository;
+import roomescape.theme.application.exception.ThemeInUseException;
 import roomescape.theme.domain.exception.ThemeNotFoundException;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeRepository;
@@ -21,12 +23,16 @@ public class ThemeService {
     private static final int THEME_SIZE_LIMIT = 10;
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
     public Theme addTheme(Theme theme) {
         return themeRepository.save(theme);
     }
 
     public void deleteTheme(Long id) {
+        if (reservationRepository.existsByTheme(id)) {
+            throw new ThemeInUseException("해당 테마의 예약이 존재합니다.");
+        }
         if (themeRepository.delete(id) == NONE_EFFECTED) {
             throw new ThemeNotFoundException("존재하지 않는 테마입니다.");
         }
